cpp06 ex01 typedef reviewed (probably mac os specific fix), ex02: cast by ref refactored

+++ cpp06/ex01/main.hpp
@@ -8,7 +8,7 @@ struct Data {
 	int age;
 };
 
-typedef unsigned long long uintptr_t;
+typedef unsigned long uintptr_t;
 
 uintptr_t serialize(const Data* ptr);
 Data* deserialize(const uintptr_t raw);
